pipeline/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
httpx==0.28.1
river>=0.21.0
typer>=0.12.0
pyyaml==6.0.2
google-cloud-firestore>=2.16.0
